refactor: align package structure with domain concepts
