+++ src/kabigon/core/errors.py
