--- src/kabigon/domain/models.py
@@ -1,7 +1,0 @@
-"""Compatibility-only exports for models now owned by ``kabigon.application``."""
-
-from kabigon.application.load_chain import LoadChainExplanation
-from kabigon.application.pipeline_catalog import ContentType
-from kabigon.application.pipeline_catalog import FallbackPolicy
-
-__all__ = ["ContentType", "FallbackPolicy", "LoadChainExplanation"]
