--- src/kabigon/infrastructure/__init__.py
@@ -1,13 +1,0 @@
-from .registry import LoaderDef
-from .registry import LoaderFactory
-from .registry import get_cli_loader_defs
-from .registry import get_loader_factory
-from .registry import list_loader_names
-
-__all__ = [
-    "LoaderDef",
-    "LoaderFactory",
-    "get_cli_loader_defs",
-    "get_loader_factory",
-    "list_loader_names",
-]
