--- src/kabigon/interface/__init__.py
@@ -1,5 +1,0 @@
-from .cli import app
-from .cli import main
-from .cli import run
-
-__all__ = ["app", "main", "run"]
