+++ src/kabigon/loaders/httpx.py
@@ -2,8 +2,8 @@ import logging
 
 import httpx
 
-from kabigon.domain.errors import LoaderContentError
-from kabigon.domain.loader import Loader
+from kabigon.core.errors import LoaderContentError
+from kabigon.core.loader import Loader
 
 from .utils import html_to_markdown
