+++ src/kabigon/loaders/ptt.py
@@ -1,7 +1,7 @@
 import logging
 
-from kabigon.application.source_applicability import parse_ptt_target
-from kabigon.domain.loader import Loader
+from kabigon.core.loader import Loader
+from kabigon.sources.applicability import parse_ptt_target
 
 from .httpx import HttpxLoader
