+++ src/kabigon/sources/__init__.py
@@ -1,0 +1,93 @@
+from .applicability import BBC_DOMAIN_SUFFIX
+from .applicability import CNN_DOMAIN_SUFFIX
+from .applicability import OPENAI_WEB_HOSTS
+from .applicability import PTT_HOSTS
+from .applicability import RAW_GITHUB_HOST
+from .applicability import REDDIT_DOMAINS
+from .applicability import REEL_PREFIX
+from .applicability import TRUTHSOCIAL_DOMAINS
+from .applicability import TWITTER_DOMAINS
+from .applicability import BBCTarget
+from .applicability import CNNTarget
+from .applicability import GitHubTarget
+from .applicability import NoVideoIDFoundError
+from .applicability import PDFTarget
+from .applicability import PttTarget
+from .applicability import RedditTarget
+from .applicability import ReelTarget
+from .applicability import TruthSocialTarget
+from .applicability import TwitterTarget
+from .applicability import UnsupportedURLNetlocError
+from .applicability import UnsupportedURLSchemeError
+from .applicability import VideoIDError
+from .applicability import YouTubeVideoTarget
+from .applicability import is_bbc_url
+from .applicability import is_cnn_url
+from .applicability import is_github_url
+from .applicability import is_openai_web_url
+from .applicability import is_pdf_target
+from .applicability import is_ptt_url
+from .applicability import is_reddit_url
+from .applicability import is_reel_url
+from .applicability import is_truthsocial_url
+from .applicability import is_twitter_url
+from .applicability import is_youtube_video_url
+from .applicability import parse_bbc_target
+from .applicability import parse_cnn_target
+from .applicability import parse_github_raw_content_target
+from .applicability import parse_github_target
+from .applicability import parse_pdf_target
+from .applicability import parse_ptt_target
+from .applicability import parse_reddit_target
+from .applicability import parse_reel_target
+from .applicability import parse_truthsocial_target
+from .applicability import parse_twitter_target
+from .applicability import parse_youtube_video_target
+
+__all__ = [
+    "BBC_DOMAIN_SUFFIX",
+    "CNN_DOMAIN_SUFFIX",
+    "OPENAI_WEB_HOSTS",
+    "PTT_HOSTS",
+    "RAW_GITHUB_HOST",
+    "REDDIT_DOMAINS",
+    "REEL_PREFIX",
+    "TRUTHSOCIAL_DOMAINS",
+    "TWITTER_DOMAINS",
+    "BBCTarget",
+    "CNNTarget",
+    "GitHubTarget",
+    "NoVideoIDFoundError",
+    "PDFTarget",
+    "PttTarget",
+    "RedditTarget",
+    "ReelTarget",
+    "TruthSocialTarget",
+    "TwitterTarget",
+    "UnsupportedURLNetlocError",
+    "UnsupportedURLSchemeError",
+    "VideoIDError",
+    "YouTubeVideoTarget",
+    "is_bbc_url",
+    "is_cnn_url",
+    "is_github_url",
+    "is_openai_web_url",
+    "is_pdf_target",
+    "is_ptt_url",
+    "is_reddit_url",
+    "is_reel_url",
+    "is_truthsocial_url",
+    "is_twitter_url",
+    "is_youtube_video_url",
+    "parse_bbc_target",
+    "parse_cnn_target",
+    "parse_github_raw_content_target",
+    "parse_github_target",
+    "parse_pdf_target",
+    "parse_ptt_target",
+    "parse_reddit_target",
+    "parse_reel_target",
+    "parse_truthsocial_target",
+    "parse_twitter_target",
+    "parse_youtube_video_target",
+]
